pyproject.toml
[tool.poetry]
name = "retinoblastoma-detector"
version = "0.1.0"
description = ""
authors = ["Amir Navon <amirnavon3@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.5.1"
torchvision = "^0.20.1"
torchaudio = "^2.5.1"
opencv-python = "^4.10.0.84"
mediapipe = "^0.10.18"
streamlit = "^1.40.1"
scikit-learn = "^1.5.2"
pandas = "^2.2.3"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
